backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
mistralai
python-dotenv
